Dashboard Begins

+++ src/products/dto/create-product.input.ts
@@ -1,10 +1,10 @@
-import { InputType, Int, Field } from '@nestjs/graphql';
+import { InputType, Int, Field, Float } from '@nestjs/graphql';
 import { TagInput, TagType } from '../entities/tags.type';
 @InputType()
 export class CreateProductInput {
   @Field({ description: 'Title of the product' })
   title: string;
-  @Field(() => Int, { description: 'price of the product' })
+  @Field(() => Float, { description: 'price of the product' })
   price: number;
   @Field({ description: 'description of the product' })
   description: string;
